pyproject.toml
[project]
name = "root-repo"
version = "0.1.0"
description = "Github root repo"
authors = [
    {name = "parthsolanke",email = "parthsolanke2002@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
